+++ packages/component/src/Styles/defaultStyleOptions.js
@@ -11,6 +11,7 @@ const PADDING_REGULAR = 10;
 const DEFAULT_OPTIONS = {
   // Color and paddings
   accent: DEFAULT_ACCENT,
+  activeActivityOutlineColor: DEFAULT_SUBTLE,
   backgroundColor: 'White',
   cardEmphasisBackgroundColor: '#F0F0F0',
   paddingRegular: PADDING_REGULAR,
@@ -202,9 +203,24 @@ const DEFAULT_OPTIONS = {
   // Internal
   internalLiveRegionFadeAfter: 1000,
 
+  // Transcript
+  transcriptTerminatorBackgroundColor: '#595959',
+  transcriptTerminatorBorderRadius: 5,
+  transcriptTerminatorColor: 'White',
+  transcriptTerminatorFontSize: 12,
+
+  transcriptActivityVisualKeyboardIndicatorColor: DEFAULT_SUBTLE,
+  transcriptActivityVisualKeyboardIndicatorStyle: 'dashed',
+  transcriptActivityVisualKeyboardIndicatorWidth: 1,
+
+  transcriptVisualKeyboardIndicatorColor: 'Black',
+  transcriptVisualKeyboardIndicatorStyle: 'solid',
+  transcriptVisualKeyboardIndicatorWidth: 2,
+  
   //AdaptiveCard push button
-  cardPushButtonBackgroundColor: '#0063B1',
+  cardPushButtonBackGroundColor: '#0063B1',
   cardPushButtonTextColor: 'white'
+
 };
 
 export default DEFAULT_OPTIONS;