+++ packages/component/src/hooks/internal/useScrollRelative.js
@@ -1,0 +1,11 @@
+import { useCallback } from 'react';
+
+import useWebChatUIContext from './useWebChatUIContext';
+
+export default function useScrollRelative() {
+  const { scrollRelativeCallbacksRef } = useWebChatUIContext();
+
+  return useCallback((...args) => scrollRelativeCallbacksRef.current.forEach(callback => callback(...args)), [
+    scrollRelativeCallbacksRef
+  ]);
+}