+++ packages/testharness/src/elements/index.js
@@ -1,7 +1,9 @@
 import activities from './activities';
 import connectivityStatus from './connectivityStatus';
+import focusedActivity from './focusedActivity';
 import newMessageButton from './newMessageButton';
 import sendBoxTextBox from './sendBoxTextBox';
+import sendButton from './sendButton';
 import suggestedActions from './suggestedActions';
 import toastDismissButtons from './toastDismissButtons';
 import toasterHeader from './toasterHeader';
@@ -9,14 +11,17 @@ import toasts from './toasts';
 import transcript from './transcript';
 import transcriptList from './transcriptList';
 import transcriptScrollable from './transcriptScrollable';
+import transcriptTerminator from './transcriptTerminator';
 import typeFocusSink from './typeFocusSink';
 import typingIndicator from './typingIndicator';
 
 export {
   activities,
   connectivityStatus,
+  focusedActivity,
   newMessageButton,
   sendBoxTextBox,
+  sendButton,
   suggestedActions,
   toasterHeader,
   toastDismissButtons,
@@ -24,6 +29,7 @@ export {
   transcript,
   transcriptList,
   transcriptScrollable,
+  transcriptTerminator,
   typeFocusSink,
   typingIndicator
 };